Devide auth logic from Basicauthinterceptor
Devide authenticate logic from Basicauthinterceptor into Fileuserlist
class and introduce Certificateauthorityinterface.
